+++ src/Simplify.Repository.FluentNHibernate/CHANGELOG.md
@@ -1,8 +1,13 @@
 # Changelog
 
-## [1.4.1] - 2021-10-28
+## [1.5.0] - 2021-10-28
+
+### Added
+
+- Object with string identifier
 
 ### Dependencies
 
+- Simplify.Repository bump to 1.6
 - Simplify.DI bump to Simplify.DI 4.1.2
 - Simplify.FluentNHibernate bump to 2.5.1